Fix lr bug

+++ parallel/scheduler.py
@@ -7,6 +7,6 @@ def scheduler(
     eta_min: float = 0.0,
 ) -> torch.Tensor:
     # cosine annealing lr
-    progress = (t / total_steps).clamp(0.0, 1.0)
+    progress = (t.float() / total_steps.float()).clamp(0.0, 1.0)
     cosine = 0.5 * (1.0 + torch.cos(torch.pi * progress))
     return eta_min + (1.0 - eta_min) * cosine
